pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Sergey Golobokov <sergeygolobokov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
djoser = "^2.3.1"
celery = "^5.4.0"
redis = "^5.2.1"
tree = "^0.2.4"
flower = "^2.0.1"
pika = "^1.3.2"
django-redis = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
